+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
 
 [workspace.package]
 version = "0.1.2"
+edition = "2021"
 authors = ["Yuekai Jia <equation618@gmail.com>"]
 license = "GPL-3.0-or-later OR Apache-2.0 OR MulanPSL-2.0"
 homepage = "https://github.com/arceos-org/arceos"
@@ -22,7 +23,8 @@ categories = ["os", "no-std", "hardware-support"]
 [workspace.dependencies]
 axdriver_base = { path = "axdriver_base", version = "0.1" }
 axdriver_block = { path = "axdriver_block", version = "0.1" }
-axdriver_net = { path = "axdriver_net", version = "0.1" }
 axdriver_display = { path = "axdriver_display", version = "0.1" }
+axdriver_net = { path = "axdriver_net", version = "0.1" }
 axdriver_pci = { path = "axdriver_pci", version = "0.1" }
 axdriver_virtio = { path = "axdriver_virtio", version = "0.1" }
+log = "0.4"
