+++ axdriver_block/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "axdriver_block"
-edition = "2021"
+edition.workspace = true
 description = "Common traits and types for block storage drivers"
 documentation = "https://arceos-org.github.io/axdriver_crates/axdriver_block"
 keywords = ["arceos", "driver", "blk"]
@@ -17,6 +17,6 @@ bcm2835-sdhci = ["dep:bcm2835-sdhci"]
 default = []
 
 [dependencies]
-log = "0.4"
 axdriver_base = { workspace = true }
 bcm2835-sdhci = { git = "https://github.com/lhw2002426/bcm2835-sdhci.git", rev = "e974f16", optional = true }
+log = { workspace = true }
