+++ axdriver_block/src/bcm2835sdhci.rs
@@ -1,10 +1,10 @@
 //! SD card driver for raspi4
 
-extern crate alloc;
-use crate::BlockDriverOps;
 use axdriver_base::{BaseDriverOps, DevError, DevResult, DeviceType};
 use bcm2835_sdhci::Bcm2835SDhci::{EmmcCtl, BLOCK_SIZE};
 use bcm2835_sdhci::SDHCIError;
+
+use crate::BlockDriverOps;
 
 /// BCM2835 SDHCI driver (Raspberry Pi SD card).
 pub struct SDHCIDriver(EmmcCtl);
