+++ axdriver_display/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "axdriver_display"
-edition = "2021"
+edition.workspace = true
 description = "Common traits and types for graphics device drivers"
 documentation = "https://arceos-org.github.io/axdriver_crates/axdriver_display"
 keywords = ["arceos", "driver", "gpu", "framebuffer"]
