+++ axdriver_net/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "axdriver_net"
-edition = "2021"
-authors = ["Yuekai Jia <equation618@gmail.com>", "ChengXiang Qi <kuangjux@outlook.com>"]
+edition.workspace = true
+authors = [
+    "Yuekai Jia <equation618@gmail.com>",
+    "ChengXiang Qi <kuangjux@outlook.com>",
+]
 description = "Common traits and types for network device (NIC) drivers"
 documentation = "https://arceos-org.github.io/axdriver_crates/axdriver_net"
 keywords = ["arceos", "driver", "nic"]
@@ -17,8 +20,8 @@ ixgbe = ["dep:ixgbe-driver"]
 fxmac = ["dep:fxmac_rs"]
 
 [dependencies]
-spin = "0.9"
-log = "0.4"
 axdriver_base = { workspace = true }
-ixgbe-driver = { git = "https://github.com/KuangjuX/ixgbe-driver.git", rev = "8e5eb74", optional = true}
 fxmac_rs = { git = "https://github.com/elliott10/fxmac_rs.git", rev = "0dbc3916", optional = true }
+ixgbe-driver = { git = "https://github.com/KuangjuX/ixgbe-driver.git", rev = "8e5eb74", optional = true }
+log = { workspace = true }
+spin = "0.9"
