+++ axdriver_virtio/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "axdriver_virtio"
-edition = "2021"
+edition.workspace = true
 description = "Wrappers of some devices in the `virtio-drivers` crate, that implement traits in the `axdriver_base` series crates"
 documentation = "https://arceos-org.github.io/axdriver_crates/axdriver_virtio"
 keywords = ["arceos", "driver", "vritio"]
@@ -12,13 +12,15 @@ repository.workspace = true
 categories.workspace = true
 
 [features]
+alloc = ["virtio-drivers/alloc"]
 block = ["axdriver_block"]
-net = ["axdriver_net"]
-gpu = ["axdriver_display"]
+gpu = ["alloc", "axdriver_display"]
+net = ["alloc", "axdriver_net"]
 
 [dependencies]
 axdriver_base = { workspace = true }
 axdriver_block = { workspace = true, optional = true }
+axdriver_display = { workspace = true, optional = true }
 axdriver_net = { workspace = true, optional = true }
-axdriver_display = { workspace = true, optional = true}
-virtio-drivers = "0.7.4"
+log = { workspace = true }
+virtio-drivers = { version = "0.7.4", default-features = false }
