+++ axdriver_virtio/src/blk.rs
@@ -1,7 +1,8 @@
-use crate::as_dev_err;
 use axdriver_base::{BaseDriverOps, DevResult, DeviceType};
 use axdriver_block::BlockDriverOps;
 use virtio_drivers::{device::blk::VirtIOBlk as InnerDev, transport::Transport, Hal};
+
+use crate::as_dev_err;
 
 /// The VirtIO block device driver.
 pub struct VirtIoBlkDev<H: Hal, T: Transport> {
